feat: remove original API plugin
