--- packages/preset-dumi/src/api-parser/fixtures/expect/excludes.json
@@ -1,16 +1,0 @@
-{
-  "AwithExcludes": [
-    {
-      "identifier": "style",
-      "description": "inline styles",
-      "type": "CSSProperties"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    }
-  ]
-}
