--- packages/preset-dumi/src/api-parser/fixtures/expect/extends.json
@@ -1,27 +1,0 @@
-{
-  "B": [
-    {
-      "identifier": "isB",
-      "type": "boolean",
-      "default": "true",
-      "required": true
-    },
-    {
-      "identifier": "className",
-      "description": "extra CSS className for this component",
-      "type": "string"
-    },
-    {
-      "identifier": "style",
-      "description": "inline styles",
-      "type": "CSSProperties"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    }
-  ]
-}
