--- packages/preset-dumi/src/api-parser/fixtures/expect/forwardRef.json
@@ -1,29 +1,0 @@
-{
-  "forwardRef": [
-    {
-      "identifier": "className",
-      "description": "extra CSS className for this component",
-      "type": "string"
-    },
-    {
-      "identifier": "style",
-      "description": "inline styles",
-      "type": "CSSProperties"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    },
-    {
-      "identifier": "ref",
-      "type": "Ref<HTMLDivElement>"
-    },
-    {
-      "identifier": "key",
-      "type": "Key"
-    }
-  ]
-}
