--- packages/preset-dumi/src/api-parser/fixtures/expect/localeDescription.json
@@ -1,11 +1,0 @@
-{
-  "localeDescription": [
-    {
-      "identifier": "className",
-      "description": "default version",
-      "description.zh-CN": "中文说明",
-      "type": "string",
-      "default": "{ type: 'A', value: [{ name: 'B' }] }"
-    }
-  ]
-}
