--- packages/preset-dumi/src/api-parser/fixtures/expect/multiple.json
@@ -1,28 +1,0 @@
-{
-  "A": [
-    {
-      "identifier": "className",
-      "description": "extra CSS className for this component",
-      "type": "string"
-    },
-    {
-      "identifier": "style",
-      "description": "inline styles",
-      "type": "CSSProperties"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    }
-  ],
-  "multiple": [
-    {
-      "identifier": "isHello",
-      "type": "boolean",
-      "required": true
-    }
-  ]
-}
