--- packages/preset-dumi/src/api-parser/fixtures/expect/skipEmptyDoc.json
@@ -1,16 +1,0 @@
-{
-  "AwithEmptyDoc": [
-    {
-      "identifier": "className",
-      "description": "extra CSS className for this component",
-      "type": "string"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    }
-  ]
-}
