--- packages/preset-dumi/src/api-parser/fixtures/expect/skipNodeModules.json
@@ -1,21 +1,0 @@
-{
-  "AextendsNode": [
-    {
-      "identifier": "className",
-      "description": "extra CSS className for this component",
-      "type": "string"
-    },
-    {
-      "identifier": "style",
-      "description": "inline styles",
-      "type": "CSSProperties"
-    },
-    {
-      "identifier": "size",
-      "description": "component size",
-      "type": "\"small\" | \"large\"",
-      "default": "small",
-      "required": true
-    }
-  ]
-}
