--- packages/preset-dumi/src/api-parser/fixtures/expect/union.json
@@ -1,18 +1,0 @@
-{
-  "union": [
-    {
-      "identifier": "type",
-      "type": "\"primary\" | \"link\"",
-      "required": true
-    },
-    {
-      "identifier": "className",
-      "type": "string"
-    },
-    {
-      "identifier": "href",
-      "type": "string",
-      "required": true
-    }
-  ]
-}
