--- packages/preset-dumi/src/api-parser/fixtures/raw/guess/NestSrc/src/index.tsx
@@ -1,1 +1,0 @@
-export { default } from '../../../fc';
