--- packages/preset-dumi/src/api-parser/fixtures/raw/multiple.tsx
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-export { default as A } from './class';
-
-const Main: React.FC<{ isHello: boolean; }> = () => null;
-
-export default Main;
