+++ packages/preset-dumi/src/theme/loader.ts
@@ -63,7 +63,6 @@ const LOCAL_THEME_PATH = '.dumi/theme';
 const FALLBACK_THEME = `${THEME_PREFIX}default`;
 const REQUIRED_THEME_BUILTINS = [
   'Alert',
-  'API',
   'Badge',
   'Example',
   'Previewer',
