--- packages/preset-dumi/src/transformer/fixtures/remark-api/Hello/index.d.ts
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-export declare const Hello: React.FC<{ className?: string }>;
-
-export const World: typeof Hello;
-
-export default Hello;
