--- packages/preset-dumi/src/transformer/fixtures/remark-api/api-slugs.md
@@ -1,11 +1,0 @@
-## First
-
-first
-
-<API src="./Hello/index.tsx"></API>
-
-## Second
-
-test custom exports
-
-<API src="./Hello/index.tsx" exports='["World"]'></API>
