--- packages/preset-dumi/src/transformer/fixtures/remark-api/auto-detect/index.tsx
@@ -1,5 +1,0 @@
-import type React from 'react';
-
-const Hello: React.FC<{ className?: string }> = () => null;
-
-export default Hello;
