--- packages/preset-dumi/src/transformer/fixtures/remark-api/guess-name.md
@@ -1,2 +1,0 @@
-<API src="./Hello/World.tsx" exports='["default"]'></API>
-<API src="./Hello/index.d.ts" exports='["default"]'></API>
