--- packages/preset-dumi/src/transformer/fixtures/remark-api/packages/pkgA/README.md
@@ -1,1 +1,0 @@
-<API src="./src/index.tsx" />
